[pg_dump_anon] Consistent backups
Closes #299

We maintain an open transaction during the entire process and then use the
snapshot_id at each step of the export process either with `pg_dump` or `psql`.
At the end, the connection is closed and the snapshot is released.

The anonymous dump is consistent and based on the snapshot of the initial
transaction.

Main changes are:

* Declare `pg_dump_anon.sh` as obsolete
* Use bats for testing pg_dump_anon
* Rename the main file to main.go
* Distribute the binaries with goreleaser